+++ Holovibes/sources/gui/windows/panels/import_panel.cc
@@ -33,6 +33,8 @@ void ImportPanel::on_notify()
     const bool no_comp = api_.compute.get_is_computation_stopped();
     ui_->InputBrowseToolButton->setEnabled(no_comp);
     ui_->FileReaderProgressBar->setVisible(!no_comp && api_.input.get_import_type() == ImportType::File);
+
+    ui_->FileLoadKindComboBox->setCurrentIndex(static_cast<int>(api_.input.get_file_load_kind()));
 }
 
 void ImportPanel::load_gui(const json& j_us)
@@ -44,8 +46,7 @@ void ImportPanel::load_gui(const json& j_us)
     ui_->ImportInputFpsSpinBox->setValue(json_get_or_default(j_us, 10000, "import", "fps"));
     update_fps(); // Required as it is called `OnEditedFinished` only.
 
-    ui_->LoadFileInGpuCheckBox->setChecked(json_get_or_default(j_us, false, "import", "from gpu"));
-    ui_->LoadFileInRamCheckBox->setChecked(json_get_or_default(j_us, false, "import", "from ram"));
+    ui_->FileLoadKindComboBox->setCurrentIndex(json_get_or_default(j_us, 0, "import", "load file kind"));
 }
 
 void ImportPanel::save_gui(json& j_us)
@@ -53,8 +54,7 @@ void ImportPanel::save_gui(json& j_us)
     j_us["panels"]["import export hidden"] = ui_->ImportExportFrame->isHidden();
 
     j_us["import"]["fps"] = ui_->ImportInputFpsSpinBox->value();
-    j_us["import"]["from gpu"] = ui_->LoadFileInGpuCheckBox->isChecked();
-    j_us["import"]["from ram"] = ui_->LoadFileInRamCheckBox->isChecked();
+    j_us["import"]["load file kind"] = ui_->FileLoadKindComboBox->currentIndex();
 }
 
 std::string& ImportPanel::get_file_input_directory()
@@ -146,19 +146,7 @@ void ImportPanel::update_import_file_path()
     api_.input.set_input_file_path(ui_->ImportPathLineEdit->text().toStdString());
 }
 
-void ImportPanel::update_load_file_in_gpu(bool enabled)
-{
-    api_.input.set_load_file_in_gpu(enabled);
-    if (enabled)
-        ui_->LoadFileInRamCheckBox->setChecked(false);
-}
-
-void ImportPanel::update_load_file_in_ram(bool enabled)
-{
-    api_.input.set_load_file_in_ram(enabled);
-    if (enabled)
-        ui_->LoadFileInGpuCheckBox->setChecked(false);
-}
+void ImportPanel::update_file_load_kind(int kind) { api_.input.set_file_load_kind(static_cast<FileLoadKind>(kind)); }
 
 void ImportPanel::update_input_file_start_index()
 {